// web/src/app/core/types/button.ts
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
import { ThemePalette } from '@angular/material/core';

type ButtonAction = 'link' | 'script' | 'noop';

export interface IButton {
  name: string;
  title?: string;
  type?: ButtonAction;
  disabled?: boolean;
  color?: ThemePalette;
  icon?: string;
  hideSpinner?: () => void;
}

/**
 * @field `value: any` - transmitted data model
 * @field `button: IButton` - target event
 */
export interface EmmitData {
  value: any;
  button: IButton;
}